xen/arm: smccc: add support for SMCCCv1.2 extended input/output registers
SMCCC v1.2 in addition to v1.1 allows x8-x17 for arm64 to be used as
parameter registers and x4—x17 for arm64 r4-r7 for arm32 to be used as
result registers in SMC64 and SMC32 respectively. Arm Firmware Framework
for Armv8-A specification makes use of x0-x7 as parameter and result
registers.

Let us add new interface to support this extended set of input/output
registers namely x0-x17 for arm64 and r0-r7  for arm32 as both parameter
and result registers.

This is based on 02a5ebe284fc ("arm64: smccc: Add support for SMCCCv1.2
extended input/output registers") from the Linux kernel and patched up
to support the arm32 case too.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>