xen/arm: add FF-A mediator
Adds a FF-A [1] mediator to communicate with a Secure Partition in
secure world.

The implementation is the bare minimum to be able to communicate with
OP-TEE running as an SPMC at S-EL1.

This is loosely based on the TEE mediator framework and the OP-TEE
mediator.

Just as with the OP-TEE mediator does OP-TEE in secure world need to
know when a guest is created and destroyed. Or more exactly OP-TEE need
to be prepared to be able to receive the first request and when
resources can be freed because the last request is received.

With the OP-TEE mediator there's a special ABI for this, but there is no
such ABI in FF-A so as a temporary hack faked forwarded rxtx map and
unmap requests are used to signal guest creation and destruction
respectively. It is expected that the FF-A specification will be
extended with a special ABI to handle this case.

[1] https://developer.arm.com/documentation/den0077/latest
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>